back/requirements.txt
django==2.2.8
djangorestframework
django-cors-headers
django-admin-honeypot
dnspython
Pillow
djongo
djangorestframework-jwt
